refactor: MAT-223 불필요한 SidebarContext 제거
