+++ src/components/Layout/Layout.css.ts
@@ -1,5 +1,8 @@
 import { style } from '@vanilla-extract/css';
 
+import { NAVBAR_HEIGHT } from '@/constants';
+import { lightThemeVars } from '@/styles/theme.css';
+
 export const layoutContainer = style({
   display: 'flex',
   width: '100%',
@@ -19,4 +22,11 @@ export const contentArea = style({
   backgroundColor: '#f5f5f5',
   padding: 16,
   overflow: 'auto',
+});
+
+export const contentAreaNested = style({
+  marginLeft: 'auto',
+  backgroundColor: lightThemeVars.color.white.hover,
+  paddingTop: NAVBAR_HEIGHT,
+  minHeight: '100vh',
 });
