+++ src/components/Navbar/Navbar.tsx
@@ -1,10 +1,10 @@
-import { useSidebar } from '@/components/Sidebar';
-
 import * as styles from './Navbar.css';
 
-export function Navbar() {
-  const { width } = useSidebar();
+interface NavbarProps {
+  width: number;
+}
 
+export function Navbar({ width }: NavbarProps) {
   return (
     <header className={styles.navbar} style={{ left: `${width}px` }}>
       <h1 className={styles.title}>교내 리그 3R 매치 기록</h1>
