+++ src/components/Sidebar/index.ts
@@ -1,4 +1,3 @@
 export * from './Sidebar';
-export * from './SidebarContext';
 export * from './NavItem';
 export * from './FooterItem';
